Add shake effect
Introduce a new `shake` function to apply a shaking effect to emote images, generating looping GIFs with adjustable intensity and damping simulation.

+++ emote/utils/effects.py
@@ -568,7 +568,7 @@ async def invert(emote: Emote) -> Emote:
     return emote
 
 
-async def shake(emote: Emote, intensity: int = 1) -> Emote:
+async def shake(emote: Emote, intensity: float = 1) -> Emote:
     """
         Applies a shaking effect to the emote image data by creating a looping shaking GIF.
